add ability to add custom widget in settings.py
Currently if you use the widget in Django admin you will notice that when using mobile, the widget does not display correctly. I have to customise the widget to fix this but I have to do it with a custom widget.
The change will allow adding a custom widget if necessary by defining CKEDITOR_5_CUSTOM_WIDGET in settings.py in your project.